Add IPv6 support
This commit modifies the "family" passed to `Addrinfo` to be
"AF_UNSPEC", which, as per
http://man7.org/linux/man-pages/man3/getaddrinfo.3.html outlines:

> The value AF_UNSPEC indicates that
> getaddrinfo() should return socket addresses for any
> address family (either IPv4 or IPv6, for example) that
> can be used with node and service.

So that we can support both IPv4 and IPv6 addresses.

Additionally, this commit modifies the way that the hostname is
determined from the middleware. By using `URI::HTTP.hostname` instead of
`URI::HTTP.host`, we can correctly determine the hostname of a URI
formatted like `http://[::1]`.